[Gold IV] Title: 좋다, Time: 16 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/1253. 좋다/좋다.cc
@@ -29,7 +29,7 @@ int main(void){
     int res = 0;
 
     rep(i, 0, n){
-        long target = vec[i]; // 판단할 수
+        int target = vec[i]; // 판단할 수
         int left = 0;
         int right = n - 1;
